Refactor: Move transformer Estimator-only code to r1 folder.
PiperOrigin-RevId: 286103069
