+++ official/r1/transformer/transformer.py
@@ -24,10 +24,10 @@ from __future__ import print_function
 
 import tensorflow as tf  # pylint: disable=g-bad-import-order
 
-from official.transformer.model import attention_layer
+from official.r1.transformer import attention_layer
+from official.r1.transformer import embedding_layer
+from official.r1.transformer import ffn_layer
 from official.transformer.model import beam_search
-from official.transformer.model import embedding_layer
-from official.transformer.model import ffn_layer
 from official.transformer.model import model_utils
 from official.transformer.utils.tokenizer import EOS_ID
